reina/requirement.txt
pandas==2.2.0
openai==1.75.0
google-genai==1.8.0
python-dotenv==1.1.0
tqdm==4.67.1
requests==2.32.3
